+++ package/kedro_viz/server.py
@@ -1,21 +1,76 @@
 """`kedro_viz.server` provides utilities to launch a webserver
 for Kedro pipeline visualisation."""
 
-import multiprocessing
 from pathlib import Path
 from typing import Any, Dict, Optional
 
-import fsspec
-import uvicorn
-from watchgod import RegExpWatcher, run_process
+from kedro.framework.session.store import BaseSessionStore
+from kedro.io import DataCatalog
+from kedro.pipeline import Pipeline
 
-from kedro_viz.api import apps
 from kedro_viz.api.rest.responses import save_api_responses_to_fs
 from kedro_viz.constants import DEFAULT_HOST, DEFAULT_PORT
+from kedro_viz.data_access import DataAccessManager, data_access_manager
+from kedro_viz.database import make_db_session_factory
+from kedro_viz.integrations.kedro import data_loader as kedro_data_loader
+from kedro_viz.integrations.kedro.sqlite_store import SQLiteStore
 from kedro_viz.launchers.utils import _check_viz_up, _wait_for
-from kedro_viz.load_data import load_and_populate_data
 
 DEV_PORT = 4142
+
+
+def populate_data(
+    data_access_manager: DataAccessManager,
+    catalog: DataCatalog,
+    pipelines: Dict[str, Pipeline],
+    session_store: BaseSessionStore,
+    stats_dict: Dict,
+):  # pylint: disable=redefined-outer-name
+    """Populate data repositories. Should be called once on application start
+    if creating an api app from project.
+    """
+
+    if isinstance(session_store, SQLiteStore):
+        session_store.sync()
+        session_class = make_db_session_factory(session_store.location)
+        data_access_manager.set_db_session(session_class)
+
+    data_access_manager.add_catalog(catalog, pipelines)
+
+    # add dataset stats before adding pipelines as the data nodes
+    # need stats information and they are created during add_pipelines
+    data_access_manager.add_dataset_stats(stats_dict)
+
+    data_access_manager.add_pipelines(pipelines)
+
+
+def load_and_populate_data(
+    path: Path,
+    env: Optional[str] = None,
+    include_hooks: bool = False,
+    package_name: Optional[str] = None,
+    pipeline_name: Optional[str] = None,
+    extra_params: Optional[Dict[str, Any]] = None,
+):
+    """Loads underlying Kedro project data and populates Kedro Viz Repositories"""
+
+    # Loads data from underlying Kedro Project
+    catalog, pipelines, session_store, stats_dict = kedro_data_loader.load_data(
+        path,
+        env,
+        include_hooks,
+        package_name,
+        extra_params,
+    )
+
+    pipelines = (
+        pipelines
+        if pipeline_name is None
+        else {pipeline_name: pipelines[pipeline_name]}
+    )
+
+    # Creates data repositories which are used by Kedro Viz Backend APIs
+    populate_data(data_access_manager, catalog, pipelines, session_store, stats_dict)
 
 
 def run_server(
@@ -53,6 +108,14 @@ def run_server(
             take precedence over) the parameters retrieved from the project
             configuration.
     """
+    # Importing below dependencies inside `run_server` to avoid ImportError when calling `load_and_populate_data` from VSCode
+    import multiprocessing
+
+    import fsspec
+    import uvicorn
+    from watchgod import RegExpWatcher, run_process
+
+    from kedro_viz.api import apps
 
     path = Path(project_path) if project_path else Path.cwd()
